Improve CPU transcription quality: int16 compute type and separate hotwords from vocabulary
- Upgrade CPU compute type from int8 to int16 for better accuracy
- Split vocabulary hints into two distinct mechanisms: vocabulary terms
  (initial_prompt context conditioning) and hotwords (decoder token bias)
- Add dedicated hotwords UI with warning about hallucination risk
- Keep hotwords small (under ~10 terms) while vocabulary terms can be large

+++ whisper_loader.py
@@ -30,7 +30,7 @@ def load_whisper_model(model_name, progress_callback=None):
     from faster_whisper import WhisperModel
 
     device = "cuda" if torch.cuda.is_available() else "cpu"
-    compute_type = "float16" if device == "cuda" else "int8"
+    compute_type = "float16" if device == "cuda" else "int16"
 
     if progress_callback:
         progress_callback(0.0)
